## Summary

This PR fixes 18 ESLint errors that were preventing the lint task from passing. The changes modernize the codebase to use current JavaScript/TypeScript best practices and improve build performance through Turbo caching.

### Key improvements:
- ✅ All ESLint errors resolved - CI now passes
- 🚀 Added Turbo caching for lint task - faster subsequent runs
- 🧹 Removed anti-pattern barrel file
- 🔒 Improved type safety with modern JavaScript methods

## Detailed Changes

### 1. ESLint Unicorn Rule Fixes
Fixed violations of the following rules across 16 files:
- `prefer-number-properties`: Use `Number.isNaN()` instead of global `isNaN()`
- `no-array-callback-reference`: Wrap callbacks in arrow functions
- `prefer-array-some`: Use `.some()` instead of `\!\!.find()`
- `no-process-exit`: Throw errors instead of `process.exit()`
- `no-useless-switch-case`: Remove redundant switch cases
- `prefer-code-point`: Use `codePointAt()` instead of `charCodeAt()`

### 2. Barrel File Removal
- Deleted `token.transaction.schema.ts` which was just re-exporting
- Updated 12 files to import directly from source
- Follows project convention against barrel files

### 3. TypeScript Compatibility
- Fixed `findLast` usage which requires ES2023
- Replaced with ES2022-compatible implementation

### 4. Turbo Configuration
Added lint task configuration in `kit/dapp/turbo.json`:
```json
{
  "lint": {
    "inputs": ["src/**/*.{js,jsx,ts,tsx}", ".eslintrc*", "eslint.config.*", "package.json", "tsconfig.json"],
    "outputs": [".eslintcache"]
  }
}
```

## Test Plan

- [x] `bun run lint` - passes with no errors
- [x] `bun run ci` - full CI suite passes
- [x] Turbo cache verified - second lint run shows "cache hit"
- [x] No breaking changes - all tests pass

## Performance Impact

Before:
- First lint run: ~1.5s
- Second lint run: ~1.5s

After:
- First lint run: ~834ms
- Second lint run: ~64ms (FULL TURBO 🚀)

## Breaking Changes

None. All changes are internal code quality improvements.

## Summary by Sourcery

Resolve ESLint unicorn rule violations across the codebase and modernize ESLint configuration for performance, while adding Turbo caching to speed up linting and updating import patterns and tests to align with new best practices.

New Features:
- Add Turbo caching configuration to the lint task for faster subsequent runs

Enhancements:
- Overhaul ESLint config: expand ignore patterns, apply targeted plugin rules, and optimize parser options for performance
- Remove barrel file and update imports to follow project convention
- Refactor code to satisfy unicorn rules (use Number.isNaN, .replaceAll, .at, spread syntax, codePointAt, switch statements, Set, numeric separators, modern JS/TS features)
- Improve type safety with updated imports (e.g., `node:crypto`), error handling (throw instead of process.exit), and modern Node/TS methods

Build:
- Update lint script to enable ESLint cache and include cache location in package.json

Tests:
- Update tests to use numeric separators for readability and adapt assertions to new parsing logic

<!-- This is an auto-generated comment: release notes by coderabbit.ai -->
## Summary by CodeRabbit

* **Chores**
  * Improved code readability and consistency by updating numeric literals to use underscore separators and refining array and string handling methods throughout the app.
  * Enhanced ESLint configuration for better performance, selective rule application, and expanded ignore patterns.
  * Updated linting scripts and task automation for improved developer workflow and caching.
  * Refined import statements and internal code structure for clarity and maintainability.
  * Added a new lint task to the build configuration for streamlined linting processes.
  * Replaced dynamic schema construction with explicit static schema definitions for token roles and actions.

* **Style**
  * Applied minor formatting changes and modernized syntax (e.g., use of `.at()`, `replaceAll`, spread operators, and `globalThis` references) for improved code clarity.
  * Converted implicit string coercions to explicit conversions for consistent UI rendering and template literals.

* **Bug Fixes**
  * Corrected pluralization logic in result count displays.
  * Improved precision in type and value checks (e.g., using `Number.isNaN`, `Number.parseFloat`, and `Number.isFinite`).

* **Refactor**
  * Simplified and clarified conditional logic and array manipulations in multiple components without altering functionality.
  * Consolidated schema imports and removed redundant files for token mutation outputs.
  * Streamlined asynchronous operations by removing unnecessary Promise wrappers.
  * Reorganized breadcrumb rendering logic for improved type safety and control flow.
  * Enhanced accordion toggle behavior for wizard sidebar groups.
  * Improved error message formatting with explicit string conversions.
  * Refined role and permission initialization logic for token middleware.

* **Tests**
  * Updated test files for improved numeric literal readability and minor syntax enhancements.

* **Documentation**
  * Added and refined inline comments and ESLint directives for code clarity.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->